created a package, moved files with servlets to this package
