+++ ProfitSite/src/main/java/Profit/PageServlet.java
@@ -1,3 +1,5 @@
+package Profit;
+
 import javax.servlet.*;
 import javax.servlet.http.*;
 import java.io.IOException;
