+++ ProfitSite/src/main/java/Profit/StatusServlet.java
@@ -1,20 +1,18 @@
+package Profit;
+
 import javax.servlet.*;
 import javax.servlet.http.*;
 import java.io.IOException;
 import java.io.PrintWriter;
 
-public class GetNameServlet extends HttpServlet {
+public class StatusServlet extends HttpServlet {
+    private static final String STATUS = "Application Is Running";
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/html");
-        String name = request.getParameter("name");
         PrintWriter pw = response.getWriter();
-        if (name.length() > 100) {
-            throw new IOException("Name is larger than 100 symbols!");
-        }
         pw.println("<html><body>");
-        pw.printf("<h1> Name is %s</h1>%n", name);
+        pw.printf("<h1 style='color:red'> %s </h1>%n", STATUS);
         pw.println("</body></html>");
     }
-
 }
