+++ ProfitSite/src/main/java/Profit/WebFilter.java
@@ -1,3 +1,5 @@
+package Profit;
+
 import javax.servlet.*;
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
